Second instance should be called SHADOW not STAGEING
Term development, stageing and production imply that there is a
roll-out relationship between these. There is none. Any developer
need to plan and execute carefully their own roll-out according
to how their themes and plugins save data in the database.

There is no one way to roll form development to production
in WordPress.

The idea of calling these SHADOW is that each shadow is a copy
of the production site at the time they are created. Developers
can then play around and experiment with the shadow and dispose
it when done.